partition worker logic into storage and compute
