+++ src/dataflow/src/sink/avro_ocf.rs
@@ -44,7 +44,7 @@ where
 
     fn render_continuous_sink(
         &self,
-        _compute_state: &mut crate::render::ComputeState,
+        _compute_state: &mut crate::server::ComputeState,
         _sink: &SinkDesc,
         sink_id: GlobalId,
         sinked_collection: Collection<G, (Option<Row>, Option<Row>), Diff>,
