+++ src/dataflow/src/sink/tail.rs
@@ -52,7 +52,7 @@ where
 
     fn render_continuous_sink(
         &self,
-        compute_state: &mut crate::render::ComputeState,
+        compute_state: &mut crate::server::ComputeState,
         sink: &SinkDesc,
         sink_id: GlobalId,
         sinked_collection: Collection<G, (Option<Row>, Option<Row>), Diff>,
